requirements.txt
# AI and LLM
openai>=1.40.0
anthropic>=0.34.0

# Discord Integration
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0
discord-ext-voice-recv>=0.2.0

# Audio Processing
SpeechRecognition>=3.10.0
pydub>=0.25.1
soundfile>=0.12.1
numpy>=1.24.0
pyaudio>=0.2.14

# TTS Options
elevenlabs>=1.5.0
gtts>=2.5.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Voice Activity Detection
webrtcvad>=2.0.10
